+++ gwcelery/tasks/superevents.py
@@ -58,8 +58,7 @@ def handle(payload):
     if alert_type == 'label_added':
         label = payload['data']['name']
         group = payload['object']['group'].lower()
-        if label == 'EM_COINC':
-            # FIXME: Add RAVEN preliminary alert
+        if label == 'RAVEN_ALERT':
             log.info('Label %s added to %s', label, gid)
         elif label not in REQUIRED_LABELS_BY_GROUP[group]:
             return
@@ -410,7 +409,9 @@ def _should_publish(event):
     trials_factor = app.conf['preliminary_alert_trials_factor'][group]
     far_threshold = app.conf['preliminary_alert_far_threshold'][group]
     far = trials_factor * event['far']
-    return not event['offline'], far <= far_threshold
+    raven_coincidence = ('RAVEN_ALERT' in event['labels'])
+
+    return not event['offline'], (far <= far_threshold or raven_coincidence)
 
 
 def keyfunc(event):